Extract shared PushNotifications test setup into a trait
