+++ plugins/woocommerce/tests/php/src/Internal/PushNotifications/Controllers/NotificationPreferencesRestControllerTest.php
@@ -4,15 +4,10 @@ declare( strict_types = 1 );
 
 namespace Automattic\WooCommerce\Tests\Internal\PushNotifications\Controllers;
 
-use Automattic\Jetpack\Connection\Manager as JetpackConnectionManager;
-use Automattic\WooCommerce\Internal\Features\FeaturesController;
 use Automattic\WooCommerce\Internal\PushNotifications\Controllers\NotificationPreferencesRestController;
 use Automattic\WooCommerce\Internal\PushNotifications\Controllers\PushTokenRestController;
-use Automattic\WooCommerce\Internal\PushNotifications\PushNotifications;
 use Automattic\WooCommerce\Internal\PushNotifications\Services\NotificationPreferencesService;
-use Automattic\WooCommerce\Proxies\LegacyProxy;
-use PHPUnit\Framework\MockObject\MockObject;
-use ReflectionClass;
+use Automattic\WooCommerce\Tests\Internal\PushNotifications\Helpers\PushNotificationsTestTrait;
 use WC_Data_Exception;
 use WC_REST_Unit_Test_Case;
 use WP_Http;
@@ -24,6 +19,8 @@ use WP_REST_Request;
  * @package WooCommerce\Tests\PushNotifications
  */
 class NotificationPreferencesRestControllerTest extends WC_REST_Unit_Test_Case {
+	use PushNotificationsTestTrait;
+
 	/**
 	 * Shop manager user ID for testing.
 	 *
@@ -37,16 +34,6 @@ class NotificationPreferencesRestControllerTest extends WC_REST_Unit_Test_Case {
 	 * @var int
 	 */
 	private $subscriber_id;
-
-	/**
-	 * @var JetpackConnectionManager|MockObject
-	 */
-	private $jetpack_connection_manager_mock;
-
-	/**
-	 * @var FeaturesController|MockObject
-	 */
-	private $features_controller_mock;
 
 	/**
 	 * Set up test.
@@ -269,62 +256,5 @@ class NotificationPreferencesRestControllerTest extends WC_REST_Unit_Test_Case {
 
 		$this->assertContains( NotificationPreferencesRestController::class, $registered_classes );
 		$this->assertContains( PushTokenRestController::class, $registered_classes );
-	}
-
-	/**
-	 * Mocks the JetpackConnectionManager class to report whether Jetpack is connected.
-	 *
-	 * @param bool $is_connected Whether the manager should report Jetpack is connected.
-	 */
-	private function mock_jetpack_connection_manager_is_connected( bool $is_connected = true ) {
-		$this->jetpack_connection_manager_mock = $this
-			->getMockBuilder( JetpackConnectionManager::class )
-			->disableOriginalConstructor()
-			->onlyMethods( array( 'is_connected' ) )
-			->getMock();
-
-		wc_get_container()->get( LegacyProxy::class )->register_class_mocks(
-			array( JetpackConnectionManager::class => $this->jetpack_connection_manager_mock )
-		);
-
-		$this->jetpack_connection_manager_mock
-			->expects( $this->any() )
-			->method( 'is_connected' )
-			->willReturn( $is_connected );
-
-		$this->reset_push_notifications_cache();
-	}
-
-	/**
-	 * Sets up the FeaturesController mock to enable the push_notifications feature.
-	 */
-	private function set_up_features_controller_mock() {
-		$this->features_controller_mock = $this
-			->getMockBuilder( FeaturesController::class )
-			->disableOriginalConstructor()
-			->onlyMethods( array( 'feature_is_enabled' ) )
-			->getMock();
-
-		$this->features_controller_mock
-			->method( 'feature_is_enabled' )
-			->willReturnCallback(
-				function ( $feature_id ) {
-					return PushNotifications::FEATURE_NAME === $feature_id;
-				}
-			);
-
-		wc_get_container()->replace( FeaturesController::class, $this->features_controller_mock );
-	}
-
-	/**
-	 * Resets the cached enablement state on the container's PushNotifications instance.
-	 */
-	private function reset_push_notifications_cache() {
-		$push_notifications = wc_get_container()->get( PushNotifications::class );
-		$reflection         = new ReflectionClass( $push_notifications );
-		$property           = $reflection->getProperty( 'enabled' );
-
-		$property->setAccessible( true );
-		$property->setValue( $push_notifications, null );
 	}
 }
